- Problem with cover item add to db solved. It has  a problem with automatic incrementation of primary keys while inserting new row for byte(tinyint). After a lot of probes various solutions id type changed into smallint.

+++ Models/Cover.cs
@@ -12,7 +12,7 @@ namespace Books.Models
         // Set primary Key as TinyInt in database
         // -----------------------------
         [Key]
-        public byte Id { get; set; }
+        public Int16 Id { get; set; }
 
         // Set Name not nullable with length 30
         // -----------------------------
